+++ backend/src/ir/instruction.hpp
@@ -88,7 +88,7 @@ namespace ir {
   std::ostream &operator<< (std::ostream &out, AddressSpace addrSpace);
 
   /*! A label is identified with an unsigned short */
-  TYPE_SAFE(LabelIndex, uint16_t)
+  TYPE_SAFE(LabelIndex, uint32_t)
 
   /*! Function class contains the register file and the register tuple. Any
    *  information related to the registers may therefore require a function
